Renamed files .jl and added function that accept a starting DateTime
